requirements.txt
fastapi
uvicorn
requests
sentence-transformers
chromadb
streamlit
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
tqdm>=4.66.0
langchain-chroma 
langchain-huggingface 
langchain-text-splitters
ollama==0.3.0
langchain-ollama